It works, without the user interrupt button

+++ src/boot/ksdk1.1.0/devSSD1331.c
@@ -180,7 +180,7 @@ devSSD1331init(void)
 }
 
 int
-drawGraph(double velocity[], int num_points)		// Will likely need to pass in velocity data here as an argument. Remember to update .h too.
+drawGraph(float velocity[], int num_points)		// Will likely need to pass in velocity data here as an argument. Remember to update .h too.
 {
 
 	/*	Clear Screen	*/
@@ -210,7 +210,7 @@ drawGraph(double velocity[], int num_points)		// Will likely need to pass in vel
 
 		
 	int hor_scale_factor = 95/(num_points-1);	
-	double max_value = 0;
+	float max_value = 0;
 	for(int i=0;i<num_points;i++)
 	{
 		if(velocity[i]>max_value)
@@ -219,15 +219,15 @@ drawGraph(double velocity[], int num_points)		// Will likely need to pass in vel
 		}
 	}
 	
-	double vert_sf = 63/max_value;
+	float vert_sf = 63/max_value;
 	
 	for(int i=0;i<num_points;i++)
 	{
 		writeCommand(kSSD1331CommandDRAWLINE);
 		writeCommand(i*hor_scale_factor);			// Start column
-		writeCommand((int)63-vert_sf*velocity[i]);		// Start row
+		writeCommand((int)(63-vert_sf*velocity[i]));		// Start row
 		writeCommand(i*hor_scale_factor);			// End column
-		writeCommand((int)63-vert_sf*velocity[i]);		// End row
+		writeCommand((int)(63-vert_sf*velocity[i]));		// End row
 		writeCommand(0xFF);					// Red
 		writeCommand(0x00);					// Green
 		writeCommand(0x00);					// Blue
